UX fixes. UI gives feedback when no comments or no bugs are retrieved. Confirms deletion of bug. Page title updated to Team 3. Page not found shown when invalid url is given.

+++ src/app/bugs-list/bugs-list.component.ts
@@ -146,6 +146,11 @@ export class BugsListComponent implements OnInit {
   }
 
   deleteRowBug(bugID: string) {
+
+    let confirm = window.confirm('You are about to delete a bug. This action is irreversible. Do you want to proceed?');
+
+    if(!confirm) return;
+
     this.restService.deleteBug(bugID).subscribe((item) => {
       this.getAllBugs();
     });
